Criando teste unitário da ConverterFiltro2CriteriaCommand
