+++ composer.json
@@ -24,7 +24,6 @@
     }
   },
   "require-dev": {
-    "phpunit/phpunit": "7.5.x-dev",
-    "paulthebaud/phpunit-generator": "^2.1"
+    "phpunit/phpunit": "7.5.x-dev"
   }
 }
